+++ lib/capistrano-redmine-deployment.rb
@@ -3,6 +3,7 @@
 require_relative 'capistrano/redmine/deployment/version'
 require_relative 'capistrano/redmine/deployment/config'
 require_relative 'capistrano/redmine/deployment/client'
+require_relative 'capistrano/redmine/deployment/railtie' if defined?(Rails)
 
 module Capistrano
   module Redmine