+++ docs/CHANGELOG.md
@@ -3,6 +3,8 @@
 ## [1.0.0] - 2024-10-01
 * **[add]** config, client & receipts
 * **[add]** rake task for setup
+* **[add]** railties setup task
+* **[fix]** deployment bugs
 
 ## [0.1.0] - 2024-09-29
 * Initial commit